Separate out use client logic into a component
